wip: move localState to agent/local
This moves the code for local agent state into the agent/local package.

The localState becomes NodeState and the separate maps for parts
of the check and service state are captured in the CheckState
and ServiceState structs.

The AntiEntropy code should also be decoupled since its function
is to sync a remote state with the local state but I didn't get to
that yet.

The code compiles but the tests do not yet.